+++ core/src/main/java/de/markusbordihn/adaptiveperformancetweakscore/config/CommonConfig.java
@@ -27,8 +27,11 @@ import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 import net.minecraftforge.fml.config.ModConfig;
+import net.minecraftforge.fml.loading.FMLPaths;
+import net.minecraftforge.fml.loading.FileUtils;
 
 import de.markusbordihn.adaptiveperformancetweakscore.Constants;
+import de.markusbordihn.adaptiveperformancetweakscore.CoreConstants;
 
 @EventBusSubscriber(bus = EventBusSubscriber.Bus.MOD)
 public final class CommonConfig {
@@ -47,7 +50,10 @@ public final class CommonConfig {
     commonSpec = specPair.getRight();
     COMMON = specPair.getLeft();
     log.info("{} common config ...", Constants.LOG_REGISTER_PREFIX);
-    ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, commonSpec);
+    FileUtils.getOrCreateDirectory(FMLPaths.CONFIGDIR.get().resolve(CoreConstants.CONFIG_ID),
+        CoreConstants.CONFIG_ID);
+    ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, commonSpec,
+        CoreConstants.CONFIG_ID_PREFIX + "core.toml");
   }
 
   public static class Config {